Make Closets Less Tanky Than Gun Safes (#35671)
* Reduce ClosetBase's DamageTrigger to 50 from 150.

* Increased LockerBase's DamageTrigger to 75, doubled LockerBaseSecure's DamageTrigger

This and the previous commit aim to improve health scaling of storage closets with their intended level of security. The
damage scale is as follows:

Inheritors of ClosetBase:       50 (old 150)
Inheritors of LockerBase:       75 (old 50)
Inheritors of LockerBaseSecure: 150 (old 75)

I left suit lockers alone because visually they look the tankiest of the lockers, though it may make sense to have them
in between secure lockers and regular lockers.

* COMMIT

---------

Co-authored-by: Princess Cheeseballs <66055347+Pronana@users.noreply.github.com>